refactor crn jobs schedule system
