+++ package.json
@@ -11,6 +11,9 @@
   "dependencies": {
     "axios": "^0.22.0",
     "bcryptjs": "^2.4.3",
+    "bree": "^9.0.0",
+    "bull": "^4.12.1",
+    "cabin": "^13.2.5",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
