+++ server.js
@@ -3,15 +3,16 @@ const fetch = require("node-fetch");
 const app = express();
 const mongoose = require("mongoose");
 const cors = require("cors");
-const { jwtDecode } = require("jwt-decode");
 const nodemailer = require("nodemailer");
 require("dotenv").config();
 const User = require("./models/user");
 const Lyrics = require("./models/lyrics");
 const SplittedLyrics = require("./models/splittedSong");
-
-
-const { createToken, hashPassword, verifyPassword } = require("./utils");
+const cron = require("node-cron");
+const Cabin = require("cabin");
+const searchRoutes = require("./routes/searchRoute");
+const autRoutes = require("./routes/authRoute");
+const songRoutes = require("./routes/songRoute");
 
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
@@ -32,89 +33,46 @@ mongoose
   .catch((err) => console.log(err.message));
 // Schedule a job to send verses at the specified frequency
 
-const scheduleJob = async (userEmail, frequency, songTitle, _id) => {
+async function processAndSendEmails() {
+  const users = await SplittedLyrics.find().distinct("userEmail");
 
-  // Limit frequency to max 3 days
-  const freq = Math.min(frequency, 3);
-  
-  let index = 0;
+  for (const userEmail of users) {
+    const userSongs = await SplittedLyrics.find({ userEmail });
 
-  const scheduleVerse = async () => {
+    for (const song of userSongs) {
+      const partOfTheSong = song.songSplitted.length;
+      const lyricsToSend = song.songSplitted.shift(); // Get and remove the first element
 
-    try {
-      
-      const user = await SplittedLyrics.findOne({userEmail});
-
-      if (user) {
-
-        const savedSongs = await SplittedLyrics.find({ 
-          $and: [
-            {_id},
-            {userEmail}  
-          ]
-        }).select("songSplitted");
-
-        if (savedSongs.length > 0) {
-          
-          const verses = savedSongs[0].songSplitted;
-          
-          if (index < verses.length) {
-
-            const unsentVerse = verses[index];
-
-            // Send verse
-            sendVerseByEmail(
-              userEmail, 
-              unsentVerse,
-              songTitle,
-              (error, response) => {
-                if (error) {
-                  console.error(error);
-                } else {
-                  console.log(response);
-                }
-              }
-            );
-
-            // Update lastSent
-            await SplittedLyrics.findOneAndUpdate(
-              {userEmail}, 
-              {lastSent: unsentVerse}, 
-              {new: true}
-            );
-            
-            index++;
-
-            // Schedule next verse
-            setTimeout(scheduleVerse, freq * 60 * 1000); 
-
-          } else {
-
-            // Finished sending verses
-            console.log("All verses sent");
-            
-            // Delete doc
-            await SplittedLyrics.findByIdAndDelete(_id);
-
-          }
-
-        }
-
+      if (lyricsToSend) {
+        await sendVerseByEmail(
+          userEmail,
+          song.songTitle,
+          lyricsToSend,
+          partOfTheSong
+        );
+        song.lastSent = new Date().toISOString();
+        await song.save();
+      } else {
+        console.log(
+          `No more lyrics to send for ${song.songTitle}. Removing from the database.`
+        );
+        await SplittedLyrics.findByIdAndRemove(song._id);
       }
-      
-    } catch (error) {
-      console.error(error);
     }
+  }
+}
 
-  };
+cron.schedule(`*/1 * * * *  `, async () => {
+  console.log("Running the cron job");
+  await processAndSendEmails();
+});
 
-  // Schedule first verse
-  setTimeout(scheduleVerse, freq * 60 * 1000); 
-
-};
-
-
-const sendVerseByEmail = (userEmail, unsentVerse, songTitle) => {
+const sendVerseByEmail = (
+  userEmail,
+  songTitle,
+  lyricsToSend,
+  partOfTheSong
+) => {
   const transporter = nodemailer.createTransport({
     service: "gmail",
     host: "smtp.gmail.com",
@@ -132,7 +90,7 @@ const sendVerseByEmail = (userEmail, unsentVerse, songTitle) => {
       address: process.env.GMAIL_USER,
     },
     to: userEmail,
-    subject: "Schedule Email",
+    subject: `Schedule Email ${songTitle} part ${partOfTheSong}`,
     html: `<div
     style="
     font-family:monospace;
@@ -142,7 +100,7 @@ const sendVerseByEmail = (userEmail, unsentVerse, songTitle) => {
     line-height: 2">
     <h2>${songTitle}</h2>
     <h3></h3>
-    ${unsentVerse}
+    ${lyricsToSend}
     </div>`,
   };
 
@@ -181,300 +139,18 @@ app.get("/v.1/api/all/:email", async (req, res) => {
   res.json(allSongs);
 });
 
-app.get("/v.1/api/song/:id", async (req, res) => {
-  const { id } = req.params;
-  const song = await Lyrics.findById(id);
-  res.json(song);
-});
+app.use(songRoutes);
+app.use(searchRoutes);
+app.use(autRoutes);
 
-app.get("/v.1/api/:selectParam/:artist", async (req, res) => {
-  try {
-    const api_key = process.env.VITE_API_KEY_MUSICMATCH;
-    const { artist, selectParam } = req.params;
-    const baseUrl = "https://api.musixmatch.com/ws/1.1/track.search";
-    const queryParams = `?${selectParam}=${artist}&page_size=4&page=1&f_has_lyrics=1&s_track_rating=desc&apikey=${api_key}`;
-    const api_url = `${baseUrl}${queryParams}`;
-    const fetch_results = await fetch(api_url);
-    const json = await fetch_results.json();
-    const result = json.message.body.track_list;
-    res.send(result);
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    res.status(500).send("Internal Server Error");
-  }
-});
-
-app.get("/v.1/api/cover/2.0/:albumName", async (req, res) => {
-  const apy_key_lastfm = "d6a6878d30433cedd1a96ed2ed43eef2";
-  const { albumName } = req.params;
-  let name = albumName.replace(/ /gi, "%20");
-  const api_url = `http://ws.audioscrobbler.com/2.0/?method=album.search&album=${name}&api_key=${apy_key_lastfm}&format=json`;
-  const fetch_results = await fetch(api_url);
-  const json = await fetch_results.json();
-  const albumCover = JSON.stringify(
-    json.results.albummatches.album[0]?.image[3]["#text"]
-  );
-  res.send(albumCover);
-});
-
-app.get(
-  "/v.1/api/songs/:trackId/:songTrack/:idAlbum/:album",
-  async (req, res) => {
-    try {
-      const { trackId, songTrack, idAlbum, album } = req.params;
-      const api_key_musicmatch = process.env.VITE_API_KEY_MUSICMATCH;
-      const api_key_lastfm = process.env.VITE_API_KEY_LASTFM;
-
-      await Promise.all([
-        fetch(
-          `https://api.musixmatch.com/ws/1.1/track.lyrics.get?track_id=${trackId}&apikey=${api_key_musicmatch}`
-        ),
-        fetch(
-          `https://api.musixmatch.com/ws/1.1/track.search?q_track=${songTrack}&apikey=${api_key_musicmatch}`
-        ),
-        fetch(
-          `https://api.musixmatch.com/ws/1.1/album.tracks.get?album_id=${idAlbum}&apikey=${api_key_musicmatch}`
-        ),
-        fetch(
-          `http://ws.audioscrobbler.com/2.0/?method=album.search&album=${album}&api_key=${api_key_lastfm}&format=json`
-        ),
-      ])
-        .then((res) => Promise.all(res.map((res) => res.json())))
-        .then((data) => {
-          res.send({
-            data,
-          });
-        });
-    } catch (error) {
-      console.error("Error fetching data:", error);
-      res.status(500).json({
-        success: false,
-        message: "Internal Server Error",
-      });
-    }
-  }
-);
-
-app.get("/v.1/api/albumTrack/:idTrack/:idAlbum", async (req, res) => {
-  try {
-    const { idTrack, idAlbum } = req.params;
-    const api_key_musicmatch = process.env.VITE_API_KEY_MUSICMATCH;
-
-    await Promise.all([
-      fetch(
-        `https://api.musixmatch.com/ws/1.1/track.lyrics.get?track_id=${idTrack}&apikey=${api_key_musicmatch}`
-      ),
-      fetch(
-        `https://api.musixmatch.com/ws/1.1/album.tracks.get?album_id=${idAlbum}&apikey=${api_key_musicmatch}`
-      ),
-    ])
-      .then((res) => Promise.all(res.map((res) => res.json())))
-      .then((data) => {
-        console.log(data);
-        res.send({
-          data,
-        });
-      });
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    res.status(500).json({
-      success: false,
-      message: "Internal Server Error",
-    });
-  }
-});
-
-app.post("/v.1/api/authenticate", async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    const user = await User.findOne({
-      email,
-    }).lean();
-
-    if (!user) {
-      return res.status(403).json({
-        message: "Wrong email or password.",
-      });
-    }
-
-    const passwordValid = await verifyPassword(password, user.password);
-
-    if (passwordValid) {
-      const { password, bio, ...rest } = user;
-      const userInfo = Object.assign({}, { ...rest });
-
-      const token = createToken(userInfo);
-
-      const decodedToken = jwtDecode(token);
-      const expiresAt = decodedToken.exp;
-
-      res.json({
-        message: "Authentication successful!",
-        token,
-        userInfo,
-        expiresAt,
-      });
-    } else {
-      res.status(403).json({
-        message: "Wrong email or password.",
-      });
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(400).json({ message: "Something went wrong." });
-  }
-});
-
-app.post("/v.1/api/signup", async (req, res) => {
-  try {
-    const { email, firstName, lastName, nickName } = req.body;
-    console.log(req.body);
-
-    const hashedPassword = await hashPassword(req.body.password);
-
-    const userData = {
-      email: email.toLowerCase(),
-      nickName,
-      firstName,
-      lastName,
-      password: hashedPassword,
-      dataSaved: Date.now(),
-    };
-
-    const existingEmail = await User.findOne({
-      email: userData.email,
-    }).lean();
-
-    if (existingEmail) {
-      return res.status(400).json({ message: "Email already exists" });
-    }
-
-    const newUser = new User(userData);
-    const savedUser = await newUser.save();
-
-    if (savedUser) {
-      const token = createToken(savedUser);
-      const decodedToken = jwtDecode(token);
-      const expiresAt = decodedToken.exp;
-
-      const { nickName, firstName, lastName, email, dataSaved } = savedUser;
-      console.log(savedUser);
-
-      const userInfo = {
-        nickName,
-        firstName,
-        lastName,
-        email,
-        dataSaved,
-      };
-
-      return res.json({
-        message: "User created!",
-        token,
-        userInfo,
-        expiresAt,
-      });
-    } else {
-      return res.status(400).json(
-        {
-          message: "There was a problem creating your account",
-        },
-        console.log()
-      );
-    }
-  } catch (err) {
-    return res.status(400).json({
-      message: "There was a problem creating your account",
-    });
-  }
-});
-
-app.post("/v.1/api/song", async (req, res) => {
-  const {
-    words,
-    trackId,
-    songTitle,
-    album_id,
-    album,
-    artistName,
-    artistId,
-    userEmail,
-  } = req.body;
-  const userInfo = await User.find({ email: userEmail });
-  const userId = userInfo[0]._id;
-  const userIdExist = await Lyrics.exists({
-    $and: [{ trackId: trackId }, { _user: userId }],
-  });
-
-  if (userIdExist) {
-    res.json({
-      type: "EXIST",
-      id: "exist",
-      _id: userId,
-      message: `${songTitle} already exist in the db`,
-    });
-  } else {
-    const newSong = new Lyrics({
-      album_id,
-      album,
-      trackId,
-      artistName,
-      artistId,
-      songTitle,
-      words,
-      dataSaved: Date.now(),
-      _user: userId,
-    });
-    await newSong.save();
-    res.json({
-      type: "SUCCESS",
-      id: "saved",
-      _id: userId,
-      message: `${songTitle} has been successfully added to the db`,
-    });
-  }
-});
-
-app.post("/v.1/api/schedule/:frequency", async (req, res) => {
-  const { frequency, lyrics, userEmail, _id, songTitle } = req.body;
-  const theEnd = ["I hope you enjoyed this way of learning"];
-  const splittedLyricsArray = lyrics.split("\n\n").map((verse) => verse);
-  splittedLyricsArray.pop();
-  const songSplitted = [...splittedLyricsArray, ...theEnd];
-  const userIdExist = await SplittedLyrics.exists({ _id: _id });
-
-  if (userIdExist) {
-    res.status(400).send({ message: `Splitted song schedule already exist in the db.` });
-  } else {
-    const splittedSong = new SplittedLyrics({
-      userEmail,
-      frequency,
-      songTitle,
-      songSplitted,
-      _id,
-      lastSent: Date.now(),
-    });
-
-    await splittedSong
-      .save()
-      .then(() => {
-        scheduleJob(userEmail, frequency, songTitle, _id);
-        res.json({ message: `Song splitted received. You will get a verse via email every ${frequency} days. Have fun!!!` });
-      })
-      .catch((error) => {
-        console.error("Error saving user:", error);
-        res.status(500).send("Internal Server Error");
-      });
-  }
-});
-
+// delete a single song
 app.delete("/v.1/api/song/:id", async (req, res) => {
   const { id } = req.params;
   const deleteItem = await Lyrics.findByIdAndDelete(id);
   res.json(deleteItem);
 });
 
+// delete all the song list
 app.delete("/v.1/api/all/:email", async (req, res) => {
   const userInfo = await User.find({ email: req.params.email });
   const userId = userInfo[0]._id;
@@ -500,6 +176,46 @@ app.post("/v.1/api/delete", (req, res) => {
     }
   );
   res.json("song deleted");
+});
+
+app.post("/v.1/api/schedule/:frequency", async (req, res) => {
+  console.log(req.body);
+  const { frequency, lyrics, userEmail, _id, songTitle, songId } = req.body;
+  const theEnd = ["I hope you enjoyed this way of learning"];
+  const splittedLyricsArray = lyrics.split("\n\n").map((verse) => verse);
+  splittedLyricsArray.pop();
+  const songSplitted = [...splittedLyricsArray, ...theEnd];
+  const userIdExist = await SplittedLyrics.exists({ _id: _id });
+
+  if (userIdExist) {
+    res
+      .status(400)
+      .send({ message: `Splitted song schedule already exist in the db.` });
+  } else {
+    const splittedSong = new SplittedLyrics({
+      songId,
+      userEmail,
+      frequency,
+      songTitle,
+      songSplitted,
+      _id,
+      lastSent: Date.now(),
+    });
+
+    await splittedSong
+      .save()
+      .then(() => {
+        // scheduleJob(userEmail, frequency, songTitle, _id, songId);
+        res.json({
+          message: `Song splitted received. You will get a verse via email every ${frequency} days. Have fun!!!`,
+          data: splittedSong,
+        });
+      })
+      .catch((error) => {
+        console.error("Error saving user:", error);
+        res.status(500).send("Internal Server Error");
+      });
+  }
 });
 
 app.post("/v.1/api/send_email", async (req, res) => {
